refactor: 빌더 분리
