+++ tomcat/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -1,8 +1,12 @@
-package org.apache.coyote.http11;
+package nextstep.jwp.controller;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import org.apache.coyote.AbstractController;
+import org.apache.coyote.http11.httprequest.HttpRequest;
+import org.apache.coyote.http11.httpresponse.HttpResponse;
+import org.apache.coyote.http11.HttpStatus;
+import org.apache.coyote.http11.ResourceResponseHandler;
 
 import static org.apache.coyote.http11.HttpStatus.FOUND;
