+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -1,7 +1,9 @@
 package org.apache.coyote.http11;
 
 import java.net.Socket;
+import nextstep.jwp.controller.RequestHandler;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.httpresponse.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
