+++ tomcat/src/test/java/org/apache/coyote/http11/HttpResponseTest.java
@@ -1,6 +1,8 @@
 package org.apache.coyote.http11;
 
 import java.util.Map;
+import org.apache.coyote.http11.httprequest.HttpRequest;
+import org.apache.coyote.http11.httpresponse.HttpResponse;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
